+++ packages/components/vite.config.ts
@@ -16,7 +16,7 @@ export default defineConfig(({ mode }) => ({
     Inspect({
       open: true,
     }),
-    react() as PluginOption,
+    react({}) as PluginOption,
     dts({
       include: [
         "src/editor/components/Editor.tsx",
@@ -40,16 +40,24 @@ export default defineConfig(({ mode }) => ({
         ignored: ["**/node_modules/**", "**/dist/**", "**/build/**"],
       },
     },
-    ...(process.env.PROXY_API_REQUESTS
-      ? {
-          proxy: {
-            "/editor": {
-              target: "http://localhost:8080", // Backend server
-              changeOrigin: true, // Ensure the request appears to come from the frontend server
-            },
-          },
-        }
-      : {}),
+    proxy: {
+      "/editor": {
+        target: "http://localhost:8080", // Backend server
+        changeOrigin: true, // Ensure the request appears to come from the frontend server
+        secure: false,
+        configure: (proxy, _options) => {
+          proxy.on("error", (err, _req, _res) => {
+            console.log("proxy error", err);
+          });
+          proxy.on("proxyReq", (proxyReq, req, _res) => {
+            console.log("Sending Request to the Target:", req.method, req.url);
+          });
+          proxy.on("proxyRes", (proxyRes, req, _res) => {
+            console.log("Received Response from the Target:", proxyRes.statusCode, req.url);
+          });
+        },
+      },
+    },
   },
   resolve: {
     alias: {
@@ -69,17 +77,15 @@ export default defineConfig(({ mode }) => ({
         "use-page-data": "src/editor/hooks/use-page-data.ts",
         "get-theme-css": "src/shared/utils/get-theme-css.ts",
       },
+      cssFileName: "upstart-components",
       formats: ["es"],
     },
     minify: process.env.NODE_ENV === "production" && process.env.NOMINIFY !== "1",
-    // cssMinify: false,
     rollupOptions: {
       external: [
         "react-icons",
         "react",
         "react-dom",
-        // "ajv",
-        "react/jsx-runtime",
         "@sinclair/typebox",
         "@upstart.gg/style-system",
         "@upstart.gg/sdk",
@@ -87,6 +93,8 @@ export default defineConfig(({ mode }) => ({
         "lodash",
       ],
       output: {
+        preserveModules: true,
+        exports: "named",
         chunkFileNames: "chunks/[name].[hash].js",
         assetFileNames: "assets/[name][extname]",
         entryFileNames: "[name].js",