+++ packages/sdk/src/shared/datasources/fetcher.ts
@@ -1,17 +1,16 @@
 import type { PageAttributes } from "../attributes";
-import type { ProviderOptions } from "./provider-options";
 
 export type DatasourceFetcherParams<
   OAuthProps = unknown,
-  Opts extends Record<string, unknown> = ProviderOptions,
+  Opts extends Record<string, unknown> = Record<string, unknown>,
 > = {
   options: Opts;
-  pageAttributes: PageAttributes;
   oauth: OAuthProps;
+  pageAttributes?: PageAttributes;
 };
 
 export type DatasourceFetcher<
   T = unknown,
   OAuthOpts = unknown,
-  Opts extends Record<string, unknown> = ProviderOptions,
+  Opts extends Record<string, unknown> = Record<string, unknown>,
 > = (params: DatasourceFetcherParams<OAuthOpts, Opts>) => Promise<T>;