+++ packages/sdk/src/shared/themes/color-system.ts
@@ -104,24 +104,30 @@ export function generateColorsVars(theme: Theme) {
     const colorName = kebabCase(_colorName);
     // Add the original color as the default (without number suffix)
     shades[`color-${colorName}`] = color;
+    // Alias it to 500
+    shades[`color-${colorName}-500`] = color;
 
     // Generate base color content
     const contentColor = getContrastingTextColor(color);
     shades[`color-${colorName}-content`] = contentColor;
+    shades[`color-${colorName}-500-content`] = contentColor;
 
-    const darkest = chroma(color).darken(2.5);
-    const lightest = chroma(color).brighten(3.5);
+    const baseColor = chroma(color);
+    const darkest = baseColor.darken(2.5);
+    const lightest = baseColor.brighten(3.5);
 
-    const colorScale = chroma.scale([lightest, color, darkest]).domain([50, 500, 900]).mode("lch"); // Use LCH color space for better perceptual scaling
+    const colorScale = chroma.scale([lightest, baseColor, darkest]).domain([50, 500, 900]).mode("oklch");
 
     // Generate shades for other colors than base colors
     if (!colorName.startsWith("base")) {
       shadeNumbers.forEach((shade) => {
+        if (shade === 500) {
+          return;
+        }
         const varName = `color-${colorName}-${shade}`;
 
         // Gen color for the specific shade
         const shadedColor = colorScale(shade);
-        // @ts-ignore oklch is a valid color format
         shades[varName] = shadedColor.css("oklch");
 
         // Gen content color for this same shade