+++ packages/sdk/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @upstart.gg/sdk
 
+## 0.0.127
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @upstart.gg/style-system@0.0.127
+
+## 0.0.126
+
+### Patch Changes
+
+- [#405](https://github.com/upstart-gg/upstart/pull/405) [`bd2de9d`](https://github.com/upstart-gg/upstart/commit/bd2de9d763362bd1cdcc079ac1c1e0905fd78462) Thanks [@mattallty](https://github.com/mattallty)! - add language to page and site
+
+- Updated dependencies []:
+  - @upstart.gg/style-system@0.0.126
+
+## 0.0.125
+
+### Patch Changes
+
+- [#401](https://github.com/upstart-gg/upstart/pull/401) [`cc78908`](https://github.com/upstart-gg/upstart/commit/cc78908733126366e657a64e5223071f4156d373) Thanks [@mattallty](https://github.com/mattallty)! - Various changes
+
+- Updated dependencies [[`cc78908`](https://github.com/upstart-gg/upstart/commit/cc78908733126366e657a64e5223071f4156d373)]:
+  - @upstart.gg/style-system@0.0.125
+
 ## 0.0.124
 
 ### Patch Changes